backend/requirements.txt
fastapi==0.108.0
uvicorn==0.32.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pandas==2.2.1
numpy==1.26.4
opencv-python==4.11.0.86
pytesseract==0.3.13
pillow==10.2.0
pdf2image==1.17.0
transformers==4.49.0
torch==2.4.1
torchvision==0.19.1